+++ app/src/main/java/fi/tv7/taivastv7/fragments/ArchivePlayerFragment.java
@@ -72,6 +72,7 @@ import static fi.tv7.taivastv7.helpers.Constants.SLASH_WITH_SPACES;
 import static fi.tv7.taivastv7.helpers.Constants.SUBTITLES_URL;
 import static fi.tv7.taivastv7.helpers.Constants.TRANSLATION_LANG_ID;
 import static fi.tv7.taivastv7.helpers.Constants.TV_BRAND;
+import static fi.tv7.taivastv7.helpers.Constants.TV_MAIN_FRAGMENT;
 import static fi.tv7.taivastv7.helpers.Constants.VIDEO_CONTROLS_TIMEOUT;
 import static fi.tv7.taivastv7.helpers.Constants.VIDEO_POSITION_TIMEOUT;
 import static fi.tv7.taivastv7.helpers.Constants.VIDEO_SEEK_STEP_SECONDS;
@@ -270,6 +271,18 @@ public class ArchivePlayerFragment extends Fragment implements Player.EventListe
     }
 
     /**
+     * Home button pressed - release player and timers.
+     */
+    public void onHomeButtonPressed() {
+        this.releasePlayer();
+        this.cancelVideoProgressTimer();
+        this.cancelVideoControlsTimer();
+
+        sharedCacheViewModel.clearPageHistory();
+        Utils.toPage(TV_MAIN_FRAGMENT, getActivity(), true, false,null);
+    }
+
+    /**
      * Handles keydown events.
      * @param keyCode
      * @param events
@@ -362,6 +375,31 @@ public class ArchivePlayerFragment extends Fragment implements Player.EventListe
             else if (keyCode == KeyEvent.KEYCODE_DPAD_UP) {
                 if (BuildConfig.DEBUG) {
                     Log.d(LOG_TAG, "ArchivePlayerFragment.onKeyDown(): KEYCODE_DPAD_UP: keyCode: " + keyCode);
+                }
+            }
+            else if (keyCode == KeyEvent.KEYCODE_MEDIA_PAUSE) {
+                if (BuildConfig.DEBUG) {
+                    Log.d(LOG_TAG, "ArchivePlayerFragment.onKeyDown(): KEYCODE_MEDIA_PAUSE: keyCode: " + keyCode);
+                }
+
+                if (!paused) {
+                    this.showControls();
+                    this.cancelVideoControlsTimer();
+
+                    this.pause();
+                }
+            }
+            else if (keyCode == KeyEvent.KEYCODE_MEDIA_PLAY) {
+                if (BuildConfig.DEBUG) {
+                    Log.d(LOG_TAG, "ArchivePlayerFragment.onKeyDown(): KEYCODE_MEDIA_PLAY: keyCode: " + keyCode);
+                }
+
+                if (paused) {
+                    if (seeking) {
+                        this.seekTo();
+                    }
+
+                    this.play();
                 }
             }
             else if (keyCode == KeyEvent.KEYCODE_BACK) {