+++ app/src/main/java/fi/tv7/taivastv7/model/SharedCacheViewModel.java
@@ -38,7 +38,13 @@ public class SharedCacheViewModel extends ViewModel {
 
     public void setPageToHistory(String page) {
         if (pageHistory != null) {
-            this.pageHistory.push(page);
+            pageHistory.push(page);
+        }
+    }
+
+    public void clearPageHistory() {
+        if (pageHistory != null) {
+            pageHistory.clear();
         }
     }